backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
django==5.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
psycopg[binary]==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via psycopg
inflection==0.5.1
    # via djangorestframework
